pyproject.toml
[project]
name = "crosswords"
version = "0.1.0"
description = "Utilities for crossword puzzles"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bs4>=0.0.2",
    "lzstring>=1.0.4",
    "pillow>=12.0.0",
    "requests>=2.32.5",
    "unidecode>=1.4.0",
]
